# Implement stopping string search in string space