--- a/modules/text_generation.py
+++ b/modules/text_generation.py
@@ -205,13 +205,15 @@ def _generate_reply(question, state, stopping_strings=None, is_chat=False):
     if len(all_stop_strings) > 0 and not state['stream']:
         state['stream'] = True
 
-    for reply in generate_func(question, original_question, seed, state, stopping_strings, is_chat=is_chat):
+    for i, reply in enumerate(generate_func(question, original_question, seed, state, stopping_strings, is_chat=is_chat)):
         reply, stop_found = apply_stopping_strings(reply, all_stop_strings)
         if is_stream:
             cur_time = time.time()
             if cur_time - last_update > 0.041666666666666664:  # Limit streaming to 24 fps
                 last_update = cur_time
                 yield reply
+        elif i == 0:
+            yield reply
 
         if stop_found:
             break