* Fixes inconsistent behavior of stopping strings with transformers.
* Extends stopping string support for all backends.
* Simplifies the code base.

This required activating streaming in the background when stopping strings are present but streaming is disabled.

## Performance test

I have made a worst-case performance test with exllama running llama-7b-4bit-128g and the following 20 long stopping strings:

```
'KMS6pQgkfLCQqweZyiq2YYq4EGzO', 'QdvPyWGxEstCUxZElD6L1u2uRp4P', 'EreYMseEtqNm1MelOEhUXprpCeIB', 'Uk7zJGqCMJpzhCuHnJfxQmsB0guk', 'W2m5VXoZf3Kkyan4Iz1zLrIf7Jqu', 'kaLBRf18oEarXbqg3TtFjlUVQTo9', 'Sn0LWu7WtDZze6ZNnUn1qVwkoJxC', 'AUIuso0oYayGxHrpgskbacpRiY7W', 'YyYPo6aM8FvuDeWtgktUBlFrbc4C', 'hA3iEivUnmY6eiuiXlcMv1HbjUC3', 'yWMXiZXFTYO1V1kH5X0j1dZAbpYx', 'JukydBzYKWXZxF6zCu8tkaxSLKfb', 'XQmDrjzjk6WL6gETRIsKYlPgaACk', 'eiDVfRyh3tA99I6gJY0eAUmyy0EU', 'pSAXQRUa4HdDPZEqShvmvRfcwjt4', 'aH05cdV888YSiAvQtODEYXAXr9aC', 'WQHm3IwrfXJBjU5ZaFXxrJf3L7Jb', 'Kj9TbuWW8tTOE7UuPheEUSvvRNAB', '2Tcfe1MDpACccrWOWmRZDDdsMgp6', 'SxtVdMo7zJTrsHD3VQfWROlAdJeX'
```

#### With the stopping strings

Output generated in 3.58 seconds (111.88 tokens/s, 400 tokens, context 6, seed 17)
Output generated in 3.58 seconds (111.88 tokens/s, 400 tokens, context 6, seed 17)
Output generated in 3.58 seconds (111.81 tokens/s, 400 tokens, context 6, seed 17)

#### Without the stopping strings

Output generated in 3.56 seconds (112.30 tokens/s, 400 tokens, context 6, seed 17)
Output generated in 3.57 seconds (112.11 tokens/s, 400 tokens, context 6, seed 17)
Output generated in 3.56 seconds (112.22 tokens/s, 400 tokens, context 6, seed 17)

-----

So the result is that in the worst case (many tokens/second, many stopping strings, long stopping strings), this implementation causes a reduction of 0.3% in the number of tokens/second. In practice it will be less than that.