@Yiximail 2023-06-24
Hmm, the gap still exists
how much performance drop depends on cpu performance

### With the stopping strings
![image](https://github.com/oobabooga/text-generation-webui/assets/15323975/90ee6e37-edd9-4bf1-9f09-27f1d4f49319)

### Without the stopping strings
![image](https://github.com/oobabooga/text-generation-webui/assets/15323975/d1fcae5a-b2ae-457c-9152-177866904842)