@oobabooga 2023-06-24
@Yiximail how many stopping strings are you using? It's strange that the difference is so large for you. I have an i5-12400F, so the performance shouldn't be any different.