@Yiximail 2023-06-24
> @Yiximail how many stopping strings are you using? It's strange that the difference is so large for you. I have an i5-12400F, so the performance shouldn't be any different.

only one, In fact it's just the difference between having and not having

![image](https://github.com/oobabooga/text-generation-webui/assets/15323975/74e10483-7325-420c-a16d-195708060616)
![image](https://github.com/oobabooga/text-generation-webui/assets/15323975/4e21cc85-d1d8-41f8-962d-19d4fbc7f72f)

when I set the stopping strings, one of the cpu cores will be fully loaded
and the gpu usage decreases with the generation length

in my test
slowing down the gpu is not the detection of stopping strings but `get_reply_from_output_ids` (mainly the decode part)
whenever a token is generated, the entire output is going to re-decoded


You tried it in 12400f? Strange, I have a visible performance drop