@Yiximail 2023-06-24
> Is that autogptq, exllama, or llama.cpp?

here? in exllama_HF