@Yiximail 2023-06-24
My previous solution was to decode only once per frame and decode again after interruption(to prevent outputting without updating the response)
That will greatly reduce the pressure on the CPU, but not good enough.

The best approach would be to leverage other cores.
As currently, only one core is handling both iteration and decoding.
![image](https://github.com/oobabooga/text-generation-webui/assets/15323975/f0f5eb03-c32c-43a3-bff1-265906bb3b25)