@official-elinas 2023-06-30
@oobabooga Stopping strings don't work in exllama or exllama_hf with extended context.